+++ tests/test_onnx.py
@@ -1,16 +1,16 @@
 import numpy as np
 import onnxruntime as rt
+from pns.util import circular_data
 from skl2onnx import to_onnx
 
 from skpns import ExtrinsicPNS, IntrinsicPNS, InverseExtrinsicPNS
-from skpns.util import circular_data
 
 
 def test_IntrinsicPNS_onnx(tmp_path):
     path = tmp_path / "pns.onnx"
 
     pns = IntrinsicPNS()
-    X = circular_data().astype(np.float32)
+    X = circular_data([0, -1, 0]).astype(np.float32)
     Xpred = pns.fit_transform(X)
 
     onx = to_onnx(pns, X[:1])
@@ -29,7 +29,7 @@ def test_ExtrinsicPNS_onnx(tmp_path):
     path = tmp_path / "pns.onnx"
 
     pns = ExtrinsicPNS()
-    X = circular_data().astype(np.float32)
+    X = circular_data([0, -1, 0]).astype(np.float32)
     Xpred = pns.fit_transform(X)
 
     onx = to_onnx(pns, X[:1])
@@ -48,7 +48,7 @@ def test_InverseExtrinsicPNS_onnx(tmp_path):
     path = tmp_path / "pns.onnx"
 
     pns = ExtrinsicPNS()
-    X = pns.fit_transform(circular_data()).astype(np.float32)
+    X = pns.fit_transform(circular_data([0, -1, 0])).astype(np.float32)
     Xpred = pns.inverse_transform(X)
 
     onx = to_onnx(InverseExtrinsicPNS(pns), X[:1])